again2

+++ dict/migrations/0006_favorite.py
@@ -16,7 +16,7 @@ class Migration(migrations.Migration):
         migrations.CreateModel(
             name='Favorite',
             fields=[
-                ('id', models.AutoField(primary_key=True, serialize=False)),
+                ('id', models.BigIntegerField(primary_key=True, serialize=False)),
                 ('user', models.ForeignKey(on_delete=django.db.models.deletion.CASCADE, to=settings.AUTH_USER_MODEL)),
                 ('word', models.ForeignKey(on_delete=django.db.models.deletion.CASCADE, to='dict.Word')),
             ],
