+++ dict/migrations/0010_auto_20200428_0129.py
@@ -13,6 +13,6 @@ class Migration(migrations.Migration):
         migrations.AlterField(
             model_name='comment',
             name='id',
-            field=models.AutoField(primary_key=True, serialize=False),
+            field=models.BigIntegerField(primary_key=True, serialize=False),
         ),
     ]
